feat: restructure backend app
