+++ backend/pyproject.toml
@@ -22,4 +22,4 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-exclude = ["frontend"]
+where = ["app"]
